nlp/requirements.txt
praw
kafka-python
python-dotenv
pylint
black
elasticsearch
flair
termcolor
